requirements.txt
tk
cinemagoer
pytest
Pillow
requests
markdown2
tkhtmlview